error solving

+++ app/src/main/java/com/virtuali/cinebok/model/GlobalClass.java
@@ -2,6 +2,6 @@ package com.virtuali.cinebok.model;
 
 public class GlobalClass {
 
-    public static int snackTot;
+    public static double snackTot;
 
 }
